chore(api): add endpoints for pages
